+++ lib/fast_immutable_collections.dart
@@ -11,7 +11,7 @@ export "src/base/number_extension.dart";
 export "src/base/object_extensions.dart";
 export "src/base/other_extensions.dart";
 export "src/base/sort.dart";
-
+export "src/base/zip_extensions.dart";
 // IList:
 export "src/ilist/from_ilist_mixin.dart";
 export "src/ilist/from_iterable_ilist_mixin.dart";
@@ -23,17 +23,14 @@ export "src/ilist/list_extension.dart";
 export "src/ilist/modifiable_list_from_ilist.dart";
 export "src/ilist/reversed_list_view.dart";
 export "src/ilist/unmodifiable_list_from_ilist.dart";
-
 // IMap:
 export "src/imap/entry.dart";
 export "src/imap/imap.dart";
 export "src/imap/map_extension.dart";
 export "src/imap/modifiable_map_from_imap.dart";
 export "src/imap/unmodifiable_map_from_imap.dart";
-
 // IMapOfSets:
 export "src/imap_of_sets/imap_of_sets.dart";
-
 // ISet:
 export "src/iset/from_iset_mixin.dart";
 export "src/iset/from_iterable_iset_mixin.dart";
@@ -41,10 +38,8 @@ export "src/iset/iset.dart";
 export "src/iset/modifiable_set_from_iset.dart";
 export "src/iset/set_extension.dart";
 export "src/iset/unmodifiable_set_from_iset.dart";
-
 // Iterator:
 export "src/iterator/iterator_extensions.dart";
-
 // ListSet and ListMap:
 export "src/list_map/list_map.dart";
 export "src/list_map/list_map_view.dart";